+++ src/main/java/nbc/devmountain/domain/chat/model/ChatMessage.java
@@ -28,7 +28,7 @@ public class ChatMessage {
 	private ChatRoom chatRoom;
 
 	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "userId")
+	@JoinColumn(name = "user_id")
 	private User user;
 
 	private String message;
