+++ src/main/java/nbc/devmountain/domain/chat/model/chatroom/dto/request/ChatRoomRequest.java
@@ -1,0 +1,7 @@
+package nbc.devmountain.domain.chat.model.chatroom.dto.request;
+
+
+public record ChatRoomRequest(
+	String chatroomName) {
+
+}
