--- src/main/java/nbc/devmountain/domain/chatroom/dto/request/ChatRoomRequest.java
@@ -1,7 +1,0 @@
-package nbc.devmountain.domain.chatroom.dto.request;
-
-
-public record ChatRoomRequest(String chatroomName) {
-
-	// private final String room_type;
-}
